Refactor fully shard rewrite after convergence fixes
