+++ megatron/core/distributed/fsdp/src/megatron_fsdp/fully_shard_rewrite/fully_shard.py
@@ -1,37 +1,27 @@
 """
-Fully Sharded Data Parallel (FSDP) Implementation
+Public fully_shard API for the Megatron-FSDP rewrite path.
 
-This module provides a PyTorch FSDP2-compatible API layer for Megatron Core.
-It wraps modules with FSDP sharding semantics, managing parameter sharding,
-unsharding, gradient reduction, and checkpointing.
-
-Key components:
-- fully_shard(): Main API to wrap a module with FSDP
-- FSDPModule: Mixin class added to wrapped modules
-- ParameterGroup: Groups parameters with shared buffers
-- DataParallelBuffer: Flat buffer for parameter/gradient storage
-
-Sharding Strategies:
-- "no_shard": No sharding (like DDP)
-- "optim": Shard optimizer state only
-- "optim_grads": Shard gradients and optimizer state
-- "optim_grads_params": Full sharding (like ZeRO-3)
+The implementation is split across:
+- fsdp_module.py: FSDPModule behavior and runtime state
+- hooks.py: forward/backward hook registration
 """
 
-import functools
-from dataclasses import dataclass, field
-from typing import Any, Callable, Dict, List, Optional, Tuple
+from typing import Callable, Optional
 
-import torch
 import torch.nn as nn
-from torch.autograd import Variable
 from torch.distributed.device_mesh import DeviceMesh
-from torch.distributed.tensor import DTensor
 from torch.distributed.tensor.placement_types import Shard
-from torch.utils._pytree import tree_flatten, tree_map, tree_unflatten
 
-from .param_group import ParameterGroup
-from .utils import ParamGroupIdx, RegisterFSDPBackwardFunction, _replace_module_parameter
+from .fsdp_module import FSDPModule
+from .hooks import (
+    _register_backward_hook,
+    _register_backward_pre_hook,
+    _register_forward_hook,
+    _register_forward_pre_hook,
+)
+from .mixed_precision import FullyShardMixedPrecisionPolicy
+
+__all__ = ["FSDPModule", "fully_shard"]
 
 
 def fully_shard(
@@ -42,7 +32,7 @@ def fully_shard(
     shard_placement_fn: Optional[
         Callable[[nn.Parameter], Optional[Shard]]
     ] = None,  # TODO: implement
-    mp_policy: Optional["MixedPrecisionPolicy"] = None,  # TODO: implement
+    mp_policy: Optional[FullyShardMixedPrecisionPolicy] = None,  # TODO: implement
     offload_policy: Optional["OffloadPolicy"] = None,  # TODO: implement
     ignored_params: Optional[set[nn.Parameter]] = None,
     # --- Megatron-FSDP specific options ---
@@ -59,37 +49,28 @@ def fully_shard(
     3. Creates ParameterGroup for each group with dedicated buffers
     4. Registers forward/backward hooks for unshard/reshard/reduce
     5. Replaces module parameters with DTensor representations
-
-    Args:
-        module: The module to wrap with FSDP.
-        mesh: DeviceMesh defining the distributed topology.
-        ignored_params: Parameters to ignore during sharding.
-
-    Returns:
-        The module wrapped with FSDP (class modified to be FSDPModule).
-
-    Example:
-        >>> mesh = init_device_mesh("cuda", (8,))
-        >>> model = ToyModel(dim=512, n_layers=3)
-        >>> for layer in model.layers:
-        ...     fully_shard(layer, mesh=mesh)
-        >>> fully_shard(model, mesh=mesh)
     """
     if isinstance(module, FSDPModule):
         raise ValueError(
             "The input module has already been fully sharded. "
             "Please do not call fully_shard on the same module more than once."
         )
+    if mp_policy is None:
+        mp_policy = FullyShardMixedPrecisionPolicy(
+            main_params_dtype=None,
+            main_grads_dtype=None,
+            grad_comm_dtype=None,
+        )
 
-    # Convert module class to FSDPModule via mixin pattern.
-    # This allows isinstance checks while preserving the original class.
     cls = module.__class__
     new_cls = type(f"FSDP{cls.__name__}", (FSDPModule, cls), {})
     module.__class__ = new_cls
 
-    # Initialize FSDP state and parameter groups
     module._init_named_param_groups(
-        mesh, ignored_params, mp_policy=mp_policy, gradient_scaling_factor=gradient_scaling_factor
+        mesh,
+        ignored_params,
+        mp_policy=mp_policy,
+        gradient_scaling_factor=gradient_scaling_factor,
     )
     module._init_fsdp_state(
         enable_unshard_prefetch=enable_unshard_prefetch,
@@ -97,884 +78,11 @@ def fully_shard(
     )
     module._init_param_main_grad_func()
 
-    # Register hooks for unshard/reshard during forward/backward
     _register_forward_pre_hook(module)
     _register_forward_hook(module)
     _register_backward_pre_hook(module)
     _register_backward_hook(module)
 
-    # Initial reshard to release parameters after setup
     module.reshard()
 
     return module
-
-
-class FSDPModule(nn.Module):
-    """
-    Mixin class for FSDP-wrapped modules.
-
-    This class is dynamically added to wrapped modules and provides
-    methods for managing parameter sharding state:
-    - unshard(): All-gather parameters before forward
-    - reshard(): Release unsharded buffer after forward
-    - reduce_grad(): Reduce-scatter gradients after backward
-    """
-
-    def _init_named_param_groups(
-        self,
-        mesh: Optional[DeviceMesh],
-        ignored_params: Optional[set],
-        mp_policy: Optional["MixedPrecisionPolicy"] = None,
-        gradient_scaling_factor: Optional[float] = None,
-    ):
-        """
-        Initialize parameter groups and build param name mapping.
-
-        This method:
-        1. Collects ignored modules (nested FSDP modules)
-        2. Materializes meta modules to actual devices
-        3. Groups parameters by (device, dtype, requires_grad)
-        4. Builds parameter name to parameter mapping
-        """
-        ignored_params = ignored_params or set()
-        ignored_modules = set()
-
-        # Collect nested FSDP modules as ignored
-        for _, child in self.named_modules():
-            if child is not self and isinstance(child, FSDPModule):
-                ignored_params.update(child.parameters())
-                for child_submodule in child.modules():
-                    ignored_modules.add(child_submodule)
-
-        # Materialize meta parameters to actual device
-        self._materialize_meta_module(ignored_modules, mesh=mesh)
-
-        # Create parameter groups
-        fsdp_param_groups = _get_module_fsdp_param_groups(
-            self,
-            mesh,
-            ignored_params=ignored_params,
-            mp_policy=mp_policy,
-            gradient_scaling_factor=gradient_scaling_factor,
-        )
-        setattr(self, "_fsdp_param_groups", fsdp_param_groups)
-
-        # Build param name to param mapping for later lookup
-        param_to_name = {p: n for n, p in self.named_parameters()}
-        self._named_param_groups = []
-
-        for fsdp_param_group in fsdp_param_groups:
-            param_names = []
-            for param in fsdp_param_group.params:
-                param_name = param_to_name[param]
-                param_names.append(param_name)
-            self._named_param_groups.append((param_names, fsdp_param_group))
-
-    def _init_param_main_grad_func(self):
-        """
-        Initialize main gradient getter function for each parameter.
-
-        This creates a closure that fetches the gradient from the
-        gradient buffer when accessed. It handles both sharded and
-        unsharded gradient buffers.
-        """
-
-        def main_grad_getter(p):
-            """Get main gradient from buffer with proper offset/size."""
-            gbuf = p._gbuf
-            item_id = p._item_id
-
-            gbuf_data = gbuf.fetch_unsharded_buffer()
-            assert gbuf_data is not None
-            assert gbuf_data.numel() > 0
-
-            # Get offset and size from buffer index
-            offset, size = gbuf.buffer_index._get_item_offset(item_id)
-            grad_data = gbuf_data[offset : offset + size].view(p.shape)
-
-            return grad_data
-
-        # Attach getter to each parameter
-        for param_group in self._fsdp_param_groups:
-            for param in param_group.params:
-                setattr(param, "_gbuf", param_group.main_grad_buffer)
-                setattr(param, "_item_id", param_group.param_idx[param])
-                param.get_main_grad = main_grad_getter.__get__(param)
-
-    def _materialize_meta_module(
-        self, ignored_modules: set, mesh: Optional[DeviceMesh] = None
-    ):
-        """
-        Materialize meta parameters to actual device and initialize.
-
-        This is needed for large models that cannot fit in a single GPU.
-        Meta parameters are moved to the current device and reset.
-
-        After materialization, all full (non-DTensor) parameters under this
-        FSDP unit are broadcast from DP rank 0 within *mesh*.  This ensures
-        every DP rank starts with identical parameter values even when the
-        random-number state differs across ranks (e.g., with
-        --data-parallel-random-init or when RNG divergence occurs during
-        meta-device model construction).  The broadcast happens **before**
-        FSDP param groups are built, so each rank subsequently receives the
-        correct shard of the same full parameter.
-        """
-        materialization_device = torch.cuda.current_device()
-        for name, m in self.named_modules():
-            if m in ignored_modules:
-                continue
-            # Skip modules that don't have meta parameters
-            if all(not p.is_meta for p in m.parameters(recurse=False)):
-                continue
-
-            # Materialize only meta tensors in a module, preserving
-            # non-meta tensors that are already initialized on device.
-            m._apply(
-                lambda t: (
-                    torch.empty_like(t, device=materialization_device) if t.is_meta else t
-                ),
-                recurse=False,
-            )
-            if hasattr(m, "reset_parameters"):
-                m.reset_parameters()
-            elif hasattr(m, "_reset_parameters"):
-                m._reset_parameters()
-            else:
-                raise ValueError(
-                    f"Module {name} contains meta parameters but does not have a reset_parameters method"
-                )
-
-        if mesh is not None and mesh.size() > 1:
-            dp_group = mesh.get_group()
-            src_rank = torch.distributed.get_global_rank(dp_group, 0)
-            for param_name, param in self.named_parameters():
-                if param.is_meta:
-                    continue
-                if isinstance(param, DTensor):
-                    continue
-                torch.distributed.broadcast(
-                    param.data, src=src_rank, group=dp_group, async_op=False
-                )
-
-    def _init_fsdp_state(self, enable_unshard_prefetch, enable_async_reduce_grad):
-        """Initialize FSDP state and mark nested FSDP modules as non-root."""
-        forward_order = [child for child in self.modules() if isinstance(child, FSDPModule)]
-        root_context = _FSDPRootContext(
-            ag_stream=(
-                torch.cuda.Stream() if enable_unshard_prefetch else torch.cuda.current_stream()
-            ),
-            rs_stream=(
-                torch.cuda.Stream() if enable_async_reduce_grad else torch.cuda.current_stream()
-            ),
-            forward_order=forward_order,
-            reduce_grad_buckets={id(module): [] for module in forward_order},
-            unshard_done_events={id(module): None for module in forward_order},
-            enable_unshard_prefetch=enable_unshard_prefetch,
-            enable_async_reduce_grad=enable_async_reduce_grad,
-        )
-        setattr(self, "_fsdp_state", _FSDPState())
-        setattr(self, "_fsdp_root_context", root_context)
-        for child in self.modules():
-            if child is not self and isinstance(child, FSDPModule):
-                child._init_fsdp_state(
-                    enable_unshard_prefetch=enable_unshard_prefetch,
-                    enable_async_reduce_grad=enable_async_reduce_grad,
-                )
-                child._fsdp_state._is_root = False
-                setattr(child, "_fsdp_root_context", root_context)
-
-    def unshard(self, async_op: bool = False, bwd_pass: bool = False):
-        """
-        Unshard parameters by all-gathering from the sharded buffer.
-
-        This is called pre-forward to make parameters available for
-        computation. After unsharding, each param.data points to
-        the full (unsharded) tensor.
-        """
-        torch.cuda.nvtx.range_push("MFSDP unshard")
-        ctx = self._fsdp_root_context
-        stream = ctx.ag_stream if async_op else torch.cuda.current_stream()
-
-        if async_op:
-            # Synchronize ag_stream with current_stream to guarantee that main-stream
-            # writes to parameter data are visible before the all-gather kernel reads them
-            # on ag_stream. Without this barrier, stale or partially-written parameter
-            # shards may be gathered, causing convergence divergence.
-            stream.wait_stream(torch.cuda.current_stream())
-
-        # Unshard this module and optionally prefetch next modules in the forward/backward pass
-        if async_op:
-            prefetch_modules = self._get_prefetch_next_modules(bwd_pass=bwd_pass)
-        else:
-            prefetch_modules = []
-        for module in [self] + prefetch_modules:
-            if ctx.unshard_done_events[id(module)] is not None:
-                continue  # Skip if unshard already issued for this module
-
-            # Unshard parameters for this module
-            for param_names, param_group in module._named_param_groups:
-                # Optional NaN checking for debugging
-                if getattr(module, "_enable_nan_checks", False):
-                    for name, dist_param in zip(param_names, param_group.dist_params):
-                        assert not torch.isnan(
-                            dist_param._local_tensor
-                        ).any(), f"NaN detected in dist param for parameter {name}"
-
-                with torch.cuda.stream(stream):
-                    param_group.unshard()
-
-            # Record event to track when unshard is done for this module
-            if async_op:
-                event = stream.record_event()
-                ctx.unshard_done_events[id(module)] = event
-
-        # Ensure unshard is complete before forward
-        if ctx.unshard_done_events[id(self)] is not None:
-            ctx.unshard_done_events[id(self)].wait()
-            ctx.unshard_done_events[id(self)] = None
-
-        # Replace module parameters with unsharded versions
-        for param_names, param_group in self._named_param_groups:
-            for name, param in zip(param_names, param_group.params):
-                _replace_module_parameter(self, name, param)
-
-            # Optional NaN checking for debugging
-            if getattr(self, "_enable_nan_checks", False):
-                for name, param in zip(param_names, param_group.params):
-                    assert not torch.isnan(param).any(), f"NaN detected in parameter {name}"
-
-        torch.cuda.nvtx.range_pop()
-
-    def _get_prefetch_next_modules(self, bwd_pass: bool = False) -> List["FSDPModule"]:
-        """Prefetch the next module in the forward/backward pass."""
-        ctx = self._fsdp_root_context
-        assert self in ctx.forward_order, "Current module not found in forward module order"
-
-        if bwd_pass:
-            module_order = list(reversed(ctx.forward_order))
-        else:
-            module_order = ctx.forward_order
-
-        i = None
-        for i, module in enumerate(module_order):
-            if module is self:
-                break
-        assert i is not None, "Current module index not found in forward module order"
-        if i + 1 >= len(module_order):
-            return []  # No next module to prefetch
-
-        return [module_order[i + 1]]
-
-    def reshard(self):
-        """Reshard parameters by replacing with sharded DTensors."""
-        torch.cuda.nvtx.range_push("MFSDP reshard")
-        ctx = self._fsdp_root_context
-        for param_names, param_group in self._named_param_groups:
-            param_group.reshard()
-            for name, dist_param in zip(param_names, param_group.dist_params):
-                _replace_module_parameter(self, name, dist_param)
-        ctx.unshard_done_events[id(self)] = None  # Clear unshard event for this module
-        torch.cuda.nvtx.range_pop()
-
-    def _wait_for_previous_async_reduce_grad(self):
-        """Wait for the previous async reduce_grad to complete before reducing gradients for this module."""
-        ctx = self._fsdp_root_context
-        if ctx.enable_async_reduce_grad:
-            backward_order = list(reversed(ctx.forward_order))
-            for i, module in enumerate(backward_order):
-                if i - 2 >= 0:
-                    buckets = ctx.reduce_grad_buckets[id(backward_order[i - 2])]
-                    while len(buckets) > 0:
-                        event, param_group = buckets.pop()
-                        event.wait()
-                        param_group.release_grad_buffer()
-                if module is self:
-                    break
-
-    def reduce_grad(self, async_op: bool = False):
-        """
-        Reduce gradients across data-parallel ranks.
-
-        This is called post-backward to:
-        1. Copy gradients to main gradient buffer
-        2. Perform all-reduce or reduce-scatter
-        3. Install reduced gradients to distributed parameters
-        """
-        torch.cuda.nvtx.range_push("MFSDP reduce_grad")
-        ctx = self._fsdp_root_context
-        stream = ctx.rs_stream if async_op else torch.cuda.current_stream()
-
-        # Handle pending reduce events before this module to ensure memory is freed in a timely manner.
-        self._wait_for_previous_async_reduce_grad()
-
-        # Perform reduction for this module
-        for param_names, param_group in self._named_param_groups:
-            if not param_group.requires_grad:
-                continue
-
-            # NaN check before reduction
-            if getattr(self, "_enable_nan_checks", False):
-                for param in param_group.params:
-                    if param.grad is not None:
-                        assert not torch.isnan(param.grad).any(), "NaN in parameter grad"
-
-            # Copy .grad → main grad buffer on main stream (fast memcpy).
-            # When gradient_accumulation_fusion is active for FSDP params, the backward
-            # kernel writes directly into main_grad (weight.main_grad = get_main_grad() in
-            # layers.py) and sets grad_added_to_main_grad=True.  In that case we must NOT
-            # zero or overwrite main_grad; the dummy .grad tensor (set by layers.py to keep
-            # backprop hooks on the main thread) should simply be discarded.
-            for name, param in zip(param_names, param_group.params):
-                main_grad = param.get_main_grad()
-                if getattr(param, 'grad_added_to_main_grad', False):
-                    if param.grad is not None:
-                        del param.grad
-                elif param.grad is None:
-                    if hasattr(param, 'main_grad') and param.main_grad is not None:
-                        if param.main_grad.data_ptr() != main_grad.data_ptr():
-                            main_grad.copy_(param.main_grad.detach())
-                    else:
-                        main_grad.zero_()
-                else:
-                    main_grad.copy_(param.grad.detach())
-                    del param.grad
-
-            if async_op:
-                # ---- Overlapped path ----
-                # Switch to rs_stream for the reduce-scatter kernel
-                stream.wait_stream(torch.cuda.current_stream())
-                with torch.cuda.stream(stream):
-                    param_group.reduce_grad()
-            else:
-                # ---- Non-overlapped path ----
-                # Reduce gradients immediately and release grad buffer
-                param_group.reduce_grad()
-                param_group.release_grad_buffer()
-
-            # Install reduced gradients to distributed parameters
-            for name, param, dist_param, dist_grad in zip(
-                param_names, param_group.params, param_group.dist_params, param_group.dist_grads
-            ):
-                if param.requires_grad and dist_grad is not None:
-                    with torch.cuda.stream(stream):
-                        dist_grad = dist_grad.to(dist_param.dtype)
-                    setattr(dist_param, "grad", dist_grad)
-
-            if async_op:
-                event = stream.record_event()
-                ctx.reduce_grad_buckets[id(self)].append((event, param_group))
-
-            # NaN check after reduction
-            if getattr(self, "_enable_nan_checks", False):
-                for name, dist_grad in zip(param_names, param_group.dist_grads):
-                    if dist_grad is not None:
-                        assert not torch.isnan(
-                            dist_grad._local_tensor
-                        ).any(), f"NaN in dist grad for parameter {name}"
-
-        torch.cuda.nvtx.range_pop()
-
-    @torch.no_grad()
-    def _scale_gradients(self, scaling_factor: float):
-        """Scale gradients by a factor (e.g., for loss scaling)."""
-        ctx = self._fsdp_root_context
-        torch.cuda.current_stream().wait_stream(ctx.rs_stream)
-        for _, child in self.named_modules():
-            if not isinstance(child, FSDPModule):
-                continue
-            for param_group in child._fsdp_param_groups:
-                for dist_grad in param_group.dist_grads:
-                    if dist_grad is None:
-                        continue
-                    dist_grad._local_tensor.mul_(scaling_factor)
-
-    def _zero_grad_buffer(self):
-        """Zero the gradient buffer for all parameter groups."""
-        for _, child in self.named_modules():
-            if not isinstance(child, FSDPModule):
-                continue
-            for param_group in child._fsdp_param_groups:
-                if param_group.main_grad_buffer is not None:
-                    param_group.main_grad_buffer.data.zero_()
-                    param_group.release_grad_buffer()
-
-                for dist_param in param_group.dist_params:
-                    if dist_param.grad is not None:
-                        del dist_param.grad
-
-    def _copy_main_weights_to_model_weights(self):
-        """Copy main weight buffer to model weight buffer."""
-        for name, child in self.named_modules():
-            if not isinstance(child, FSDPModule):
-                continue
-            for param_names, param_group in child._named_param_groups:
-                if param_group.main_weight_buffer is None:
-                    continue
-                param_group.model_weight_buffer.data.copy_(param_group.main_weight_buffer.data)
-
-    def _compute_per_param_norms(self) -> Dict[str, Dict[str, float]]:
-        """
-        Compute per-parameter L2 norms for params and grads.
-
-        Returns a dict {param_name: {"param_norm": float, "grad_norm": float}}.
-        The norms are globally reduced across DP ranks.
-        """
-        results = {}
-        dp_group = None
-        for name, child in self.named_modules():
-            if not isinstance(child, FSDPModule) or child is self:
-                continue
-            for param_names, param_group in child._named_param_groups:
-                if dp_group is None and param_group.dp_group is not None:
-                    dp_group = param_group.dp_group
-                for pname, dist_param, dist_grad in zip(
-                    param_names, param_group.dist_params, param_group.dist_grads
-                ):
-                    full_name = f"{name}.{pname}" if name else pname
-                    results[full_name] = {"param_norm": 0.0, "grad_norm": 0.0}
-                    if dist_param._local_tensor.numel() > 0:
-                        results[full_name]["param_norm"] = (
-                            dist_param._local_tensor.float().norm(p=2).item() ** 2
-                        )
-                    if dist_grad is not None and dist_grad._local_tensor.numel() > 0:
-                        results[full_name]["grad_norm"] = (
-                            dist_grad._local_tensor.float().norm(p=2).item() ** 2
-                        )
-
-        if dp_group is None:
-            for pg in self._fsdp_param_groups:
-                if pg.dp_group is not None:
-                    dp_group = pg.dp_group
-                    break
-
-        if dp_group is not None:
-            for param_name in results:
-                for key in ("param_norm", "grad_norm"):
-                    t = torch.tensor([results[param_name][key]], device="cuda")
-                    torch.distributed.all_reduce(t, group=dp_group)
-                    results[param_name][key] = t.sqrt().item()
-        return results
-
-    def _log_per_param_norms(self, iteration: int, prefix: str = ""):
-        """Log per-parameter param and gradient L2 norms via print (rank 0 only)."""
-        norms = self._compute_per_param_norms()
-        if torch.distributed.get_rank() != 0:
-            return
-        for param_name in sorted(norms.keys()):
-            pn = norms[param_name]["param_norm"]
-            gn = norms[param_name]["grad_norm"]
-            print(
-                f"{prefix} iter={iteration} param={param_name} "
-                f"param_norm={pn:.6f} grad_norm={gn:.6f}"
-            )
-
-    def _log_parameter_groups(self):
-        """Compact log of FSDP parameter groups and their parameters (rewrite path)."""
-
-        def _fmt_dtype(dt: torch.dtype) -> str:
-            short = {
-                torch.float32: "fp32",
-                torch.float16: "fp16",
-                torch.bfloat16: "bf16",
-                torch.int64: "i64",
-                torch.int32: "i32",
-                torch.uint8: "u8",
-            }
-            return short.get(dt, str(dt).removeprefix("torch."))
-
-        def _elem_size(dt: torch.dtype) -> int:
-            return {
-                torch.float32: 4,
-                torch.float16: 2,
-                torch.bfloat16: 2,
-                torch.int64: 8,
-                torch.int32: 4,
-                torch.uint8: 1,
-            }.get(dt, 1)
-
-        def _mb(b: int | float) -> str:
-            return f"{b / 1_000_000:.2f} MB"
-
-        rank = torch.distributed.get_rank()
-        lines = [f"╔══ FSDP Parameter Groups (rank {rank}) ══╗"]
-        group_idx = 0
-        total_model_elems = 0
-        total_comm = 0
-        total_pad = 0
-
-        for name, child in self.named_modules():
-            if not isinstance(child, FSDPModule) or child is self:
-                continue
-            for param_names, param_group in child._named_param_groups:
-                numel = sum(p.numel() for p in param_group.params)
-                total_model_elems += numel
-                dp_sz = torch.distributed.get_world_size(param_group.dp_group)
-                dp_rk = torch.distributed.get_rank(param_group.dp_group)
-
-                # Per-buffer metrics
-                buf_entries = []
-                group_pad = 0
-                group_comm = 0
-                for buf_label, buf in [
-                    ("W", param_group.model_weight_buffer),
-                    ("MW", param_group.main_weight_buffer),
-                    ("G", param_group.main_grad_buffer),
-                ]:
-                    if buf is None:
-                        continue
-                    gsize = buf.buffer_index.bucket_meta.size
-                    esize = _elem_size(buf.dtype)
-                    group_pad += max(0, (gsize - numel)) * esize
-                    group_comm += gsize * esize
-                    dist_flag = "D" if buf.is_distributed else "R"
-                    buf_entries.append(
-                        f"{buf_label}[{_fmt_dtype(buf.dtype)}:{buf.data_size}:{dist_flag}]"
-                    )
-                total_pad += group_pad
-                total_comm += group_comm
-
-                lines.append(
-                    f"╟── {name}  (#{group_idx})  "
-                    f"dp={dp_sz}  strat={param_group.sharding_strategy}  "
-                    f"cf={param_group.chunk_size_factor}"
-                )
-                lines.append(
-                    f"║   {numel:,} elems × {_fmt_dtype(param_group.dtype)}  "
-                    f"comm={_mb(group_comm)}  pad={_mb(group_pad)}  "
-                    f"{'  '.join(buf_entries)}"
-                )
-                for pname, param in zip(param_names, param_group.params):
-                    dist_idx = param_group.param_idx.get(param)
-                    wbuf = param_group.model_weight_buffer
-                    offset_info = ""
-                    if wbuf is not None and dist_idx is not None:
-                        ii = wbuf.buffer_index.item_index_map.get(dist_idx)
-                        if ii is not None:
-                            offset_info = f"  @{ii.global_data_index:,}+{ii.size:,}"
-                    lines.append(f"║     {pname:50s} {str(tuple(param.shape)):24s}{offset_info}")
-                group_idx += 1
-
-        lines.append(
-            f"╚══ SUMMARY: {group_idx} groups  "
-            f"{total_model_elems:,} model elems  "
-            f"comm={_mb(total_comm)}  pad={_mb(total_pad)} ══╝"
-        )
-        print("\n".join(lines))
-
-    def _set_nan_check(self, enable_nan_checks: bool):
-        """Enable or disable NaN checking."""
-        for _, child in self.named_modules():
-            if not isinstance(child, FSDPModule):
-                continue
-            setattr(child, "_enable_nan_checks", enable_nan_checks)
-
-        if enable_nan_checks:
-            for name, param in self.named_parameters():
-                if isinstance(param, DTensor):
-                    param_data = param.data._local_tensor
-                else:
-                    param_data = param.data
-                assert not torch.isnan(param_data).any(), f"NaN detected in parameter {name}"
-            for child in self.modules():
-                if not isinstance(child, FSDPModule):
-                    continue
-                for param_group in child._fsdp_param_groups:
-                    for param in param_group.params:
-                        wbuf = param_group.model_weight_buffer
-                        param_data = wbuf.get_item(param_group.param_idx[param], only_shard=False)
-                        assert not torch.isnan(
-                            param_data
-                        ).any(), "NaN detected in model weight buffer"
-
-
-class _FSDPState:
-    """
-    Internal state for FSDP module tracking.
-
-    Attributes:
-        _is_root: Whether this is the root FSDP module (handles final callback).
-        _post_backward_callback_queued: Whether callback is queued for execution.
-    """
-
-    def __init__(self):
-        self._is_root = True
-        self._post_backward_callback_queued = False
-
-
-def _get_module_fsdp_param_groups(
-    module: nn.Module,
-    mesh: Optional[DeviceMesh] = None,
-    ignored_params: Optional[set[nn.Parameter]] = None,
-    mp_policy: Optional["MixedPrecisionPolicy"] = None,
-    gradient_scaling_factor: Optional[float] = None,
-) -> List[ParameterGroup]:
-    """
-    Group module parameters by (device, dtype, requires_grad) and create ParameterGroups.
-
-    Parameters are grouped because they share the same buffer management
-    and sharding strategy. Each group gets its own DataParallelBuffer.
-    """
-    param_groups = {}
-
-    for param in module.parameters():
-        if ignored_params is not None and param in ignored_params:
-            continue
-
-        # Group by (device, dtype, requires_grad)
-        param_attrs = (param.device, param.dtype, param.requires_grad)
-        if param_attrs not in param_groups:
-            param_groups[param_attrs] = []
-        param_groups[param_attrs].append(param)
-
-    # Create ParameterGroup for each group
-    fsdp_param_groups = []
-    for i, params in enumerate(param_groups.values()):
-        fsdp_param_groups.append(
-            ParameterGroup(
-                params,
-                mesh=mesh,
-                param_group_id=ParamGroupIdx(id(module), i),
-                main_params_dtype=mp_policy.main_params_dtype if mp_policy is not None else None,
-                main_grads_dtype=mp_policy.main_grads_dtype if mp_policy is not None else None,
-                gradient_scaling_factor=gradient_scaling_factor,
-            )
-        )
-
-    return fsdp_param_groups
-
-
-def _register_forward_pre_hook(module: FSDPModule):
-    """Register pre-forward hook to unshard parameters."""
-
-    def unshard_param_groups(module, *unused):
-        ctx = module._fsdp_root_context
-        module.unshard(async_op=ctx.enable_unshard_prefetch, bwd_pass=False)
-
-    module._mfsdp_forward_pre_hook = module.register_forward_pre_hook(
-        unshard_param_groups, prepend=True
-    )
-
-
-def _register_forward_hook(module: FSDPModule):
-    """Register post-forward hook to reshard parameters."""
-
-    def reshard_param_groups(module, *unused):
-        module.reshard()
-
-    module._mfsdp_forward_hook = module.register_forward_hook(reshard_param_groups)
-
-
-def _register_backward_pre_hook(module: FSDPModule):
-    """
-    Register backward pre-hook to handle gradient computation.
-
-    This uses a custom backward hook that attaches to output tensors
-    to trigger unshard at the right time during backward pass.
-    """
-
-    def create_custom_backward_hook(module: FSDPModule, custom_backward_handler: Callable):
-        """Create a custom backward hook attached to output tensors."""
-
-        @torch.compiler.disable
-        def forward_hook(_module, inputs, output):
-            # View-as to avoid output being the same tensor object
-            output = tree_map(lambda t: t.view_as(t) if torch.is_tensor(t) else t, output)
-
-            # Collect tensor outputs
-            output_list = []
-            if isinstance(output, torch.Tensor):
-                output_list = [output]
-            elif isinstance(output, (tuple, list)):
-                output_list = [t for t in output if isinstance(t, torch.Tensor)]
-
-            # Register pre-backward hook on output tensors
-            # This triggers when gradients are computed
-            torch.autograd.graph.register_multi_grad_hook(
-                output_list, lambda grads: custom_backward_handler(_module, grads), mode="any"
-            )
-            return output
-
-        return module.register_forward_hook(forward_hook)
-
-    def pre_backward_hook(module: FSDPModule, grads):
-        """Hook called before backward pass for this module."""
-        ctx = module._fsdp_root_context
-        setattr(module, "post_backward_issued", False)
-        for param_group in module._fsdp_param_groups:
-            for param in param_group.params:
-                setattr(param, "grad_added_to_main_grad", False)
-        if module._fsdp_state._is_root and not module._fsdp_state._post_backward_callback_queued:
-            _register_post_backward_final_callback(module._fsdp_state, module)
-        module.unshard(async_op=ctx.enable_unshard_prefetch, bwd_pass=True)
-
-    module._mfsdp_backward_pre_hook = create_custom_backward_hook(
-        module, custom_backward_handler=pre_backward_hook
-    )
-
-
-def _register_backward_hook(module: FSDPModule):
-    """
-    Register backward hook using autograd Function.
-
-    This inserts a RegisterFSDPBackwardFunction in the backward pass
-    that triggers reshard and reduce_grad after gradients are computed.
-    """
-
-    def post_backward(module: FSDPModule):
-        """Hook called after backward pass for this module."""
-        ctx = module._fsdp_root_context
-        module.reshard()
-        module.reduce_grad(async_op=ctx.enable_async_reduce_grad)
-        module.post_backward_issued = True
-
-    @torch.compiler.disable
-    def _register_post_backward_hook(
-        post_backward_hook: Callable,
-        module: nn.Module,
-        args: Tuple[Any, ...],
-        kwargs: Dict[str, Any],
-    ):
-        """
-        Register a post-backward hook by inserting an autograd Function.
-
-        This approach works by registering a pre-forward hook that wraps
-        input tensors in an autograd Function. The Function's backward
-        calls the post_backward_hook after gradients are computed.
-        """
-        if not torch.is_grad_enabled():
-            return args, kwargs
-
-        # Flatten args and kwargs
-        args_list, args_spec = tree_flatten(args)
-        kwargs_list, kwargs_spec = tree_flatten(kwargs)
-        args_kwargs_list = list(args_list) + list(kwargs_list)
-
-        # Filter to tensors with gradients
-        inp_tensor_indices: List[int] = []
-        inp_tensors: List[torch.Tensor] = []
-        for i, obj in enumerate(args_kwargs_list):
-            if torch.is_tensor(obj) and obj.requires_grad:
-                inp_tensor_indices.append(i)
-                inp_tensors.append(obj)
-
-        if len(inp_tensors) == 0:
-            return args, kwargs
-
-        # Wrap inputs in autograd Function
-        # The Function's backward will call post_backward_hook
-        inp_tensors = RegisterFSDPBackwardFunction.apply(
-            functools.partial(post_backward_hook, module), *inp_tensors
-        )
-
-        # Restore args and kwargs
-        for inp_tensor_idx, inp_tensor in zip(inp_tensor_indices, inp_tensors):
-            args_kwargs_list[inp_tensor_idx] = inp_tensor
-        args_list = args_kwargs_list[: len(args_list)]
-        kwargs_list = args_kwargs_list[len(args_list) :]
-        args = tree_unflatten(args_list, args_spec)
-        kwargs = tree_unflatten(kwargs_list, kwargs_spec)
-
-        return args, kwargs
-
-    module._mfsdp_backward_hook = module.register_forward_pre_hook(
-        functools.partial(_register_post_backward_hook, post_backward), with_kwargs=True
-    )
-
-
-def _register_post_backward_final_callback(state: _FSDPState, module: nn.Module) -> None:
-    """
-    Register the final callback that runs after all backward passes complete.
-
-    This is only registered by the root FSDP module to avoid duplicate
-    callbacks. It reshards all modules and reduces gradients at the end
-    of the backward pass.
-    """
-    assert state._is_root, "Only root FSDP should register post-backward callback"
-    if state._post_backward_callback_queued:
-        return
-
-    def _post_backward_final_callback(root_state: _FSDPState, root_module: nn.Module):
-        """Final callback: reshard all modules and reduce gradients."""
-        ctx = root_module._fsdp_root_context
-        stream = ctx.rs_stream
-        for module in reversed(ctx.forward_order):
-            if getattr(module, "post_backward_issued", False):
-                continue
-            module.reshard()
-            module.reduce_grad(async_op=ctx.enable_async_reduce_grad)
-        for buckets in ctx.reduce_grad_buckets.values():
-            while len(buckets) > 0:
-                event, param_group = buckets.pop()
-                event.wait()
-                param_group.release_grad_buffer()
-        torch.cuda.current_stream().wait_stream(stream)
-        root_state._post_backward_callback_queued = False
-
-    state._post_backward_callback_queued = True
-    Variable._execution_engine.queue_callback(
-        functools.partial(_post_backward_final_callback, state, module)
-    )
-
-
-@dataclass
-class _FSDPRootContext:
-    """
-    Runtime context shared across all FSDP modules within a single root.
-
-    This object coordinates CUDA streams, execution ordering, and async
-    communication overlap (all-gather / reduce-scatter) during forward
-    and backward passes.
-    """
-
-    # ------------------------------------------------------------------
-    # CUDA streams (communication overlap)
-    # ------------------------------------------------------------------
-    ag_stream: torch.cuda.Stream  # all-gather / unshard stream
-    rs_stream: torch.cuda.Stream  # reduce-scatter stream
-
-    # ------------------------------------------------------------------
-    # Forward execution ordering
-    # ------------------------------------------------------------------
-    forward_order: List[FSDPModule] = field(default_factory=list)
-    """
-    FSDP modules in actual forward execution order.
-
-    This ordering is used to:
-    - Schedule prefetching of parameters (unshard)
-    - Ensure correct overlap between compute and communication
-    """
-
-    # ------------------------------------------------------------------
-    # Unshard (all-gather) tracking
-    # ------------------------------------------------------------------
-    unshard_done_events: Dict[int, torch.cuda.Event] = field(default_factory=dict)
-    """
-    Maps module_id -> CUDA event signaling completion of parameter unshard.
-
-    Used to enforce correct dependency between all-gather and compute.
-    """
-
-    enable_unshard_prefetch: bool = True
-    """Whether to prefetch (pipeline) parameter unshard for upcoming modules."""
-
-    # ------------------------------------------------------------------
-    # Reduce-scatter (gradient sync) tracking
-    # ------------------------------------------------------------------
-    reduce_grad_buckets: Dict[int, List[Tuple[torch.cuda.Event, "ParameterGroup"]]] = field(
-        default_factory=dict
-    )
-    """
-    Maps module_id -> list of (event, parameter_group) tuples.
-
-    Each entry corresponds to a module and contains a list of:
-        (event, parameter_group)
-
-    - event: signals gradient readiness
-    - parameter_group: gradients to be reduced
-
-    This structure enables ordered overlap of backward compute and
-    gradient synchronization.
-    """
-
-    enable_async_reduce_grad: bool = True
-    """Whether to overlap gradient reduction with backward computation."""
